Split into `engo` and `core` (#283)
Split engo into engo and common

By doing that, it fixes #245. 

A generic asset loader was implemented, which works for every platform, every file. Each platform now only has to implement a openFile method to support those files. This fixes #246. Thanks to @mewmew for the initial code. 

Some spritesheet optimization was done, to ensure a Spritesheet is just one Texture on the GPU-side of things. Fixes #253. 

Shader-compilation is now being tested (except for Windows, due to limitations of AppVeyor). Fixes #280. 

The README now has a bit more accurate information on the current state, and fixed some bugs and typos (thanks @otraore). Fixes #277. 

Moved WorldBounds to common.CameraBounds to make its usage clearer. Fixes #228. 

SetBackground was moved to common, and fixes #235 in doing so. 

A little bit more documentation was added, which helps with #261. 

Thanks to everyone for the discussion on this. As well as everyone who has helped review. This was a collaborative effort. 
 :+1: 1