font atlas
draws the font atlas much more effeciently. Gets all the dimensions for the image before it creates the image, so the image size is exactly the right size. It also only ever has one image rendered, so it doesn't multiply memory used for multiple images. I also am using the draw function from image/font, so that when the truetype font library is put into image/font, it's ready to be used there instead of using the legacy draw methods from the old truetype library.