added support for asymmetric sprite sheets as well as for borders between sprites on the sprite sheet
Added support for asymmetric sprite sheets. Created a sprite struct that keeps the position and size information about each sprite. Since the symmetric is just a subset of asymmetric, all the symmetric stuff produces sprites on it's own then feeds that to the asymmetric.
Also implemented symmetric sprite sheets with borders between the sprites, which helps with issues that sprites bleed into eachother.