Add GlobalScale option
Added a GlobalScale option that'll scale the entire game by that amount.
If the Scale is set, it'll adjust mouse clicks and rendering, but if the
user is utilizing mouse positioning or a custom render system, they'll
need to be aware of if this is set and what it sets to. Default is just
engo.Point{1,1}.